Save files to a separate chunks table
Save file to separate chunks table instead of incorporating them in to
the attachments table. This makes backups a little bit easier and adds
possibility to actually save the attachments in chunks to multiple
chunks rows which would make possible to stream the files.

Currently we just save the files in to a single chunk which has obvious
memory and performance limitations